# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TCFL Portal API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/apply/birth-certificate": {
      "get": {
        "tags": [
          "apply birth certificate"
        ],
        "summary": "Birth Certificate",
        "operationId": "birth_certificate",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthCertificateState"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "apply birth certificate"
        ],
        "summary": "Confirm Birth Certificate",
        "operationId": "confirm_birth_certificate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmBirthIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthCertificateState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "apply birth certificate"
        ],
        "summary": "Upload Birth Certificate",
        "operationId": "upload_birth_certificate",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_birth_certificate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthCertificateState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/handoff/session/birth-certificate": {
      "get": {
        "tags": [
          "apply birth certificate"
        ],
        "summary": "Phone Birth Certificate",
        "operationId": "phone_birth_certificate",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthCertificateState"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "apply birth certificate"
        ],
        "summary": "Phone Confirm Birth Certificate",
        "operationId": "phone_confirm_birth_certificate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmBirthIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthCertificateState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "apply birth certificate"
        ],
        "summary": "Phone Upload Birth Certificate",
        "operationId": "phone_upload_birth_certificate",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_phone_upload_birth_certificate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthCertificateState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BirthCertificateState": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/BirthStatus"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "capture_device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Device"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "matches_id": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matches Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "document_id",
          "received_at",
          "capture_device",
          "name",
          "date_of_birth",
          "matches_id"
        ],
        "title": "BirthCertificateState"
      },
      "BirthStatus": {
        "type": "string",
        "enum": [
          "none",
          "uploaded",
          "confirmed"
        ],
        "title": "BirthStatus"
      },
...
      "Body_phone_upload_birth_certificate": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_phone_upload_birth_certificate"
      },
...
      "Body_upload_birth_certificate": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_birth_certificate"
      },
...
      "ConfirmBirthIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 3,
            "title": "Name"
          },
          "date_of_birth": {
            "type": "string",
            "format": "date",
            "title": "Date Of Birth"
          }
        },
        "type": "object",
        "required": [
          "name",
          "date_of_birth"
        ],
        "title": "ConfirmBirthIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}